+++ R/fstg02.R
@@ -12,6 +12,7 @@
 #' @param subgroups (`character`) the subgroups variable name to list baseline risk factors.
 #' @param strata_var (`character`) required if stratified analysis is performed.
 #' @param stat_var (`character`) the names of statistics to be reported in `tabulate_survival_subgroups`.
+#' @param max_colwidth (`int`) maximum width of columns. Stratification label longer than this will be truncated.
 #' @param ... Further arguments passed to `g_forest` and `extract_rsp_subgroups` (a wrapper for
 #'  `h_odds_ratio_subgroups_df` and `h_proportion_subgroups_df`). For details, see the documentation in `tern`.
 #'  Commonly used arguments include `col_symbol_size`, `col`, `vline`, `groups_lists`, `conf_level`,
@@ -21,6 +22,7 @@
 #'  * `adam_db` object must contain the table specified by `dataset` with `"PARAMCD"`, `"ARM"`,
 #'  `"AVAL"`, `"AVALU"`, `"CNSR"`, and the columns specified by `subgroups` which is denoted as
 #'  `c("SEX", "AGEGR1", "RACE")` by default.
+#'  * If the plot is too large to be rendered in the output, please refer to `FSTG01`.
 #'
 #' @return a `gTree` object.
 #'
@@ -32,10 +34,12 @@ fstg02_main <- function(adam_db,
                         subgroups = c("SEX", "AGEGR1", "RACE"),
                         strata_var = NULL,
                         stat_var = c("n_tot", "n", "median", "hr", "ci"),
+                        max_colwidth = 10,
                         ...) {
   assert_all_tablenames(adam_db, c("adsl", dataset))
   df_lbl <- paste0("adam_db$", dataset)
   assert_string(arm_var)
+  assert_int(max_colwidth)
   assert_character(subgroups, null.ok = TRUE)
   assert_character(strata_var, null.ok = TRUE)
   assert_character(stat_var, null.ok = TRUE)
@@ -68,14 +72,19 @@ fstg02_main <- function(adam_db,
     data = adam_db[[dataset]],
     ...
   )
-
+  df$survtime$subgroup <- stringr::str_trunc(df$survtime$subgroup, max_colwidth)
+  df$hr$subgroup <- stringr::str_trunc(df$hr$subgroup, max_colwidth)
   result <- basic_table() %>%
     tabulate_survival_subgroups(df, vars = stat_var, time_unit = timeunit)
-
+  cw <- pmin(propose_column_widths(result), max_colwidth + 2)
+  final_width <- stringWidth(strrep("x", cw))
   execute_with_args(
     g_forest,
     tbl = result,
-    ...
+    ...,
+    width_row_names = final_width[1],
+    width_columns = final_width[-1],
+    draw = FALSE
   )
 }