+++ NEWS.md
@@ -7,9 +7,9 @@
 * Add `FSTG02` template.
 * Update the argument `"is_event"` in `KMG01` to `"IS_EVENT"`.
 * Update the argument `"is_rsp"` in `FSTG01` to `"IS_RSP"`.
+* `FSTG01` and `FSTG02` stratification variable labels will be truncated to make the page fit.
 * Update the script for `chevron_tlg` objects. `details` is deprecated.
 * Fix issue if call `run` method with `do.call` with `verbose`.
-
 
 # chevron 0.2.1